+++ packages/bookmark/src/BookmarkItem.tsx
@@ -41,20 +41,28 @@ export const BookmarkItem: React.FC<{
 }) => {
     const path = pathFromRoot ? `${pathFromRoot}.${index}` : `${index}`;
     const defaultIsCollapsed = React.useMemo(() => shouldBeCollapsed(bookmark), [bookmark]);
+    const bookmarkExpandedMap = store.get('bookmarkExpandedMap');
     const defaultExpanded = isBookmarkExpanded
         ? isBookmarkExpanded({ bookmark, doc, depth, index })
+        : bookmarkExpandedMap.has(path)
+        ? bookmarkExpandedMap.get(path)
         : !defaultIsCollapsed;
     const [expanded, setExpanded] = React.useState(defaultExpanded);
 
     const hasSubItems = bookmark.items && bookmark.items.length > 0;
 
-    const toggleSubItems = (): void => setExpanded((expanded) => !expanded);
+    const toggleSubItems = (): void => {
+        const newState = !expanded;
+        store.updateCurrentValue('bookmarkExpandedMap', (currentValue) => currentValue.set(path, newState));
+        setExpanded(newState);
+    };
 
     const clickBookmark = (): void => {
         if (hasSubItems && bookmark.dest) {
             onJumpToDest(bookmark.dest);
         }
     };
+
     const clickItem = (): void => {
         if (!hasSubItems && bookmark.dest) {
             onJumpToDest(bookmark.dest);