+++ CHANGELOG.md
@@ -27,8 +27,8 @@ switchTrigger(Trigger.TextSelection);
 
 **Improvements**
 
+-   Keep the expanded/collapsed state of each bookmark
 -   `RenderBookmarkItemProps` includes new `path` property that indicates the path from each bookmark item to the root
-
 -   `SelectionData` provides more information about the selected text including:
 
 | Property       | Type        | Description                                  |